Program runs but progress bar is a bit messy and the buttons flashing are weird

+++ OrcMathGUI/src/myStuff/ButtonInterfaceVictor.java
@@ -8,8 +8,8 @@ import guiTeacher.interfaces.Clickable;
 public interface ButtonInterfaceVictor extends Clickable{
 
 	void setColor(Color color);
-	void setX(int x);
-	void setY(int y);
+	void setX(double x);
+	void setY(double y);
 	void setAction(Action a);
 	void light();
 	void dim();
